Update to stylguide.html. Rename js-codesamp class to js-inject-code and remove it's styling from styleguide.scss. Add the code-sample class to handle the styling. This follows our css coding style guidelines.
Rename the file to inject-code.js to make its purpose obvious. Move the line
of code that initiates highlightJS to a separate file to separate concerns.

Move datepicker class to its own component, and include it in core.
It's there as a placeholder for now. It's an id.selector, but needs to be a
class selector. Defer that until I refactor the bootstrap out of the datepicker
packages.